goes-build: Add building of Linux, U-boot, Coreboot, and initial RAM filesystems
Replace most of the functionality of the system-build makefiles.

Add building of the Linux kernel on x86_64 and ARM.

Add building of U-boot on ARM. This support is incomplete as of this commit.
We don't yet package the kernel/initramfs/DTB into a modern uimage.

Add building of Coreboot on x86_64.

Add building of initial RAM filesystems on x86_64 and ARM.

Incremental builds are *very* fast. With warm caches on my 2.6GHZ i7
with 4 cores, an incremental build takes just over one minute.

Initial checkouts are slow, because we have to build a specialty toolchain
for Coreboot, and build several Linux kernels. This could be made faster
by passing the -j option to make, but this is not implemented yet.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>