goes-build: Module and fe1 cleanups
Fix building with the -tags=plugin option. Its not presently possible
to build without FE1 sources. The -tags=plugin option determines
whether the FE1 code is built as a plugin or statically linked,
but it is needed to build goes either way.

Clean up paths to work with go modules.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>