goes-build: More fixes for fe1 plugin
Plugins need to be built from the same directory as the main
package (yeah, you gotta be kidding!). Change the build of
fe1.so to build it from "platinasystems/go". The trick is that
you can't specify a folder, you have to specify the exact go
source files, as in:
	go build -buildmode=plugin -o fe1.so main/fe1/main.go

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>