goes-build: Legacy mode now remaps new image into the old flash layout
Code growth on the BMC (with the addition of the ssh server) now
exceeds the current flash layout. Rework legacy mode to slice up the
ITB into the files expected by earlier versions of goes.

With this change, we always build new image format, but we create
the zip file using the old format. This allows for ITB files to
be up to 5MiB. Currently the image is 4.8MiB.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>