goes-build: When building root filesystem, make goes be 'init'
Instead of having 'init' as a link to 'goes', make 'goes' be a link
to 'init'. This allows us to replace the 'goes' command dynamically.
'goes' can not overwrite itself, since it is open at the time.

Groundwork for goes-recovery. The first-stage 'goes' in boot flash
replaces itself with one found on ata flash or the network.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>