goes-build: Fix race condition with incomplete dependent packages
Using sync.Once() did not protect against the case where the goroutine
was still running building the package, since the goroutine wrapper
had already exited.

Instead serialize target builds using a mutex which we wait for in
maker goroutine. This allows for maximum parallelism as some targets
on the same dependency level may still run as the ones blocked on their
dependency wait for another goroutine to finish that dependency.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>