goes-build: Make checkout and configure logic more robust
Guard against partial checkouts and make configure logic more robust.

Treat a worktree as uninitialized if it is missing the .git file,
instead of assuming that if the directory exists, its been initialized.

Treat a worktree as unconfigured if the .config file does not exist,
or if the -branch option was used to explicitly specify a branch.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>