goes-build: Don't require qspi-header-sckl00
Don't require a copy of qspi-header-sckl00 in the build directory.
Synthesize the data ourselves.

Note that these are magic numbers copied from the file we were using.
The NXP documention does not explain the fields well. The test package
ensures that we build the right thing.

For more information, see:

   http://www.nxp.com/docs/en/reference-manual/IMX6ULRM.pdf
   https://community.nxp.com/thread/461341

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>