goes-build: Add debian control file generation
Add support for a debian/control target which reads debian/control.in,
applies substitutions, and writes debian/control.

The substitution symbol #KERNELRELEASE# is replaced with the kernel
release string.

The substitution symbol #KDEB_PKGVERSION# is replaced with the kernel
debian package version.

The substitution symbol #KERNELID# is replaced with the kernel version
number.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>