u-boot-envvars: Disable asynchronous preemption in goes-bmc
We get random crashes in the ssh daemon. Debugging pointed that
the crashes occured in signal_unix.go, in this line:

	setStack := adjustSignalStack(sig, g.m, &gsignalStack)

The value of sig was SIGURG. This is the signal used by the go
runtime for asynchronous preemption. Asynchronous preemption is a new
feature in Go 1.14.

The value of g was not valid. I am going to see if I can report this
as a bug in the go runtime. In the interim, disabling asynchronous
preemption with GODEBUG=asyncpreemptoff=1 works around the problem.

For more information, read:

https://medium.com/a-journey-with-go/go-asynchronous-preemption-b5194227371c

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>