goes-build: Allow multiple versions to coexist
Add version info to goes-build so that we can have multiple
versions installed at once. This simplifies the build process
as some changes to goes-build are not backwards compatible.
Rather than introducing a lot of options for backwards compatibility,
allow the binaries to coexist, so that the build scripts can just
pick the version they need.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>